# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BoardPrefUpdate": {
        "properties": {
          "board": {
            "enum": [
              "kraft",
              "lino",
              "salvia",
              "niebla",
              "barro",
              "pizarra"
            ],
            "title": "Board",
            "type": "string"
          }
        },
        "required": [
          "board"
        ],
        "title": "BoardPrefUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListCreate": {
        "properties": {
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ListCreate",
        "type": "object"
      },
      "ListMemberBrief": {
        "description": "A member as shown on list overviews: id and display name, nothing else.\n\nDeliberately carries no email and no photo — the overview payload must not\nleak contact details of co-members.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "display_name"
        ],
        "title": "ListMemberBrief",
        "type": "object"
      },
      "ListRead": {
        "properties": {
          "board": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board"
          },
          "cart_count": {
            "default": 0,
            "title": "Cart Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "item_count": {
            "default": 0,
            "title": "Item Count",
            "type": "integer"
          },
          "members": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ListMemberBrief"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "purchased_count": {
            "default": 0,
            "title": "Purchased Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "emoji",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "ListRead",
        "type": "object"
      },
      "ListUpdate": {
        "properties": {
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ListUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CarroQueSí API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/lists": {
      "get": {
        "operationId": "get_lists_lists_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListRead"
                  },
                  "title": "Response Get Lists Lists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Lists",
        "tags": [
          "lists"
        ]
      },
      "post": {
        "operationId": "create_list_lists_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create List",
        "tags": [
          "lists"
        ]
      }
    },
    "/lists/{list_id}": {
      "delete": {
        "operationId": "delete_list_lists__list_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete List",
        "tags": [
          "lists"
        ]
      },
      "get": {
        "operationId": "get_list_lists__list_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get List",
        "tags": [
          "lists"
        ]
      },
      "patch": {
        "operationId": "update_list_lists__list_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update List",
        "tags": [
          "lists"
        ]
      }
    },
    "/lists/{list_id}/default": {
      "put": {
        "description": "Mark this list as the caller's default (for Siri), clearing any prior one.\n\nPer-user membership state — deliberately does NOT bump lists.updated_at, since\nthe flag is invisible to co-members and shouldn't trigger their polls.",
        "operationId": "set_default_list_lists__list_id__default_put",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Default List",
        "tags": [
          "lists"
        ]
      }
    },
...
    "/lists/{list_id}/prefs/board": {
      "put": {
        "description": "Pick the board this list sits on, for the caller only.\n\nPersonal presentation state — deliberately does NOT bump lists.updated_at,\nthe same rule as set_default_list: the board is invisible to co-members\nand must not trigger their polls.",
        "operationId": "set_board_pref_lists__list_id__prefs_board_put",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-is-admin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Is-Admin"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardPrefUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Board Pref",
        "tags": [
          "lists"
        ]
      }
    },
...
  }
}